Remove duplicate code in student.js (#101)
This change removes about 50 lines of code that were duplicated between student.js and run_python.js following a prior merge (#70).

The code wasn't strictly identical, as it was the result of a bad merge from a previous state of the repository. This change also corrects a number of issues on the instructor view that had previously been fixed on the student view, such as #64, #68, and #97.